refactor(vendingmachine.view.*) : 자판기에서 나오는 요청문들은 Userview에서 VendingMachineView로 이동
